+++ src/main/java/net/minecraft/block/BlockSnow.java
@@ -27,7 +27,7 @@ public class BlockSnow extends Block {
 
     protected BlockSnow() {
         super(Material.snow);
-        this.setDefaultState(this.blockState.getBaseState().withProperty(LAYERS, Integer.valueOf(1)));
+        this.setDefaultState(this.blockState.getBaseState().withProperty(LAYERS, 1));
         this.setBlockBounds(0.0F, 0.0F, 0.0F, 1.0F, 0.125F, 1.0F);
         this.setTickRandomly(true);
         this.setCreativeTab(CreativeTabs.tabDecorations);
@@ -35,11 +35,11 @@ public class BlockSnow extends Block {
     }
 
     public boolean isPassable(IBlockAccess worldIn, BlockPos pos) {
-        return worldIn.getBlockState(pos).getValue(LAYERS).intValue() < 5;
+        return worldIn.getBlockState(pos).getValue(LAYERS) < 5;
     }
 
     public AxisAlignedBB getCollisionBoundingBox(World worldIn, BlockPos pos, IBlockState state) {
-        int i = state.getValue(LAYERS).intValue() - 1;
+        int i = state.getValue(LAYERS) - 1;
         float f = 0.125F;
         return new AxisAlignedBB((double) pos.getX() + this.minX, (double) pos.getY() + this.minY, (double) pos.getZ() + this.minZ, (double) pos.getX() + this.maxX, (float) pos.getY() + (float) i * f, (double) pos.getZ() + this.maxZ);
     }
@@ -58,7 +58,7 @@ public class BlockSnow extends Block {
 
     public void setBlockBoundsBasedOnState(IBlockAccess worldIn, BlockPos pos) {
         IBlockState iblockstate = worldIn.getBlockState(pos);
-        this.getBoundsForLayers(iblockstate.getValue(LAYERS).intValue());
+        this.getBoundsForLayers(iblockstate.getValue(LAYERS));
     }
 
     protected void getBoundsForLayers(int p_150154_1_) {
@@ -68,7 +68,7 @@ public class BlockSnow extends Block {
     public boolean canPlaceBlockAt(World worldIn, BlockPos pos) {
         IBlockState iblockstate = worldIn.getBlockState(pos.down());
         Block block = iblockstate.getBlock();
-        return block != Blocks.ice && block != Blocks.packed_ice ? (block.getMaterial() == Material.leaves ? true : (block == this && iblockstate.getValue(LAYERS).intValue() >= 7 ? true : block.isOpaqueCube() && block.blockMaterial.blocksMovement())) : false;
+        return block != Blocks.ice && block != Blocks.packed_ice ? (block.getMaterial() == Material.leaves ? true : (block == this && iblockstate.getValue(LAYERS) >= 7 ? true : block.isOpaqueCube() && block.blockMaterial.blocksMovement())) : false;
     }
 
     public void onNeighborBlockChange(World worldIn, BlockPos pos, IBlockState state, Block neighborBlock) {
@@ -86,7 +86,7 @@ public class BlockSnow extends Block {
     }
 
     public void harvestBlock(World worldIn, EntityPlayer player, BlockPos pos, IBlockState state, TileEntity te) {
-        spawnAsEntity(worldIn, pos, new ItemStack(Items.snowball, state.getValue(LAYERS).intValue() + 1, 0));
+        spawnAsEntity(worldIn, pos, new ItemStack(Items.snowball, state.getValue(LAYERS) + 1, 0));
         worldIn.setBlockToAir(pos);
         player.triggerAchievement(StatList.mineBlockStatArray[Block.getIdFromBlock(this)]);
     }
@@ -111,15 +111,15 @@ public class BlockSnow extends Block {
     }
 
     public IBlockState getStateFromMeta(int meta) {
-        return this.getDefaultState().withProperty(LAYERS, Integer.valueOf((meta & 7) + 1));
+        return this.getDefaultState().withProperty(LAYERS, (meta & 7) + 1);
     }
 
     public boolean isReplaceable(World worldIn, BlockPos pos) {
-        return worldIn.getBlockState(pos).getValue(LAYERS).intValue() == 1;
+        return worldIn.getBlockState(pos).getValue(LAYERS) == 1;
     }
 
     public int getMetaFromState(IBlockState state) {
-        return state.getValue(LAYERS).intValue() - 1;
+        return state.getValue(LAYERS) - 1;
     }
 
     protected BlockState createBlockState() {
