+++ src/main/java/net/minecraft/entity/player/EntityPlayerMP.java
@@ -220,7 +220,7 @@ public class EntityPlayerMP extends EntityPlayer implements ICrafting {
             int j = 0;
 
             while (iterator.hasNext() && j < i) {
-                aint[j++] = iterator.next().intValue();
+                aint[j++] = iterator.next();
                 iterator.remove();
             }
 
@@ -393,7 +393,7 @@ public class EntityPlayerMP extends EntityPlayer implements ICrafting {
         EntityLivingBase entitylivingbase = this.getAttackingEntity();
 
         if (entitylivingbase != null) {
-            EntityList.EntityEggInfo entitylist$entityegginfo = EntityList.entityEggs.get(Integer.valueOf(EntityList.getEntityID(entitylivingbase)));
+            EntityList.EntityEggInfo entitylist$entityegginfo = EntityList.entityEggs.get(EntityList.getEntityID(entitylivingbase));
 
             if (entitylist$entityegginfo != null) {
                 this.triggerAchievement(entitylist$entityegginfo.field_151513_e);
@@ -843,7 +843,7 @@ public class EntityPlayerMP extends EntityPlayer implements ICrafting {
         this.translator = packetIn.getLang();
         this.chatVisibility = packetIn.getChatVisibility();
         this.chatColours = packetIn.isColorsEnabled();
-        this.getDataWatcher().updateObject(10, Byte.valueOf((byte) packetIn.getModelPartFlags()));
+        this.getDataWatcher().updateObject(10, (byte) packetIn.getModelPartFlags());
     }
 
     public void loadResourcePack(String url, String hash) {
@@ -866,7 +866,7 @@ public class EntityPlayerMP extends EntityPlayer implements ICrafting {
         if (p_152339_1_ instanceof EntityPlayer) {
             this.playerNetServerHandler.sendPacket(new S13PacketDestroyEntities(new int[]{p_152339_1_.getEntityId()}));
         } else {
-            this.destroyedItemsNetCache.add(Integer.valueOf(p_152339_1_.getEntityId()));
+            this.destroyedItemsNetCache.add(p_152339_1_.getEntityId());
         }
     }
